requirements.txt
fastapi==0.128.2
uvicorn[standard]==0.40.0

pandas==3.0.0
numpy==2.4.1

scikit-learn==1.8.0
joblib==1.5.3

pydantic==2.12.5
PyYAML==6.0.3
